Don't crash if temp table doesn't exist

+++ db/migrate/20110503201404_rm_first_temp_table.rb
@@ -1,6 +1,6 @@
 class RmFirstTempTable < ActiveRecord::Migration
   def self.up
-    drop_table :check_runs_before_add_index_on_check_runs
+    drop_table :check_runs_before_add_index_on_check_runs if table_exists?(:check_runs_before_add_index_on_check_runs)
   end
 
   def self.down
